+++ js/group/Index.svelte
@@ -33,7 +33,7 @@
 <style>
 	div {
 		border: var(--block-border-width) solid var(--border-color-primary);
-		background: var(--block-border-color);
+		background: var(--border-color-primary);
 		border-radius: var(--block-radius);
 		display: flex;
 		flex-direction: column;
@@ -43,6 +43,9 @@
 	div > :global(*:not(.absolute)) {
 		border: none;
 		border-radius: 0;
+	}
+	div > :global(.gr-accordion:not(.absolute)) {
+		border: none !important;
 	}
 	.hide {
 		display: none;