+++ js/tabitem/Index.svelte
@@ -22,7 +22,7 @@
 	order={gradio.props.order}
 	scale={gradio.props.scale}
 	component_id={gradio.props.component_id}
-	on:select={({ detail }) => gradio.dispatch("select", detail)}
+	onselect={(data) => gradio.dispatch("select", data)}
 >
 	<slot />
 </TabItem>